move requires_admin into utils.py, minor fixes
