+++ src/commands.py
@@ -65,12 +65,8 @@ Say
 
 Speaks a message to the specified channel as the bot
 """
+@utils.requires_admin
 async def say(message):
-    # Only allow if user has correct permissions
-    roles = [x.id for x in message.author.roles]
-    if ADMIN_ACCESS not in roles:
-        return None
-
     try:
         payload = utils.remove_command(message.content)
         channel_id = utils.get_command(payload)
@@ -96,12 +92,8 @@ Edit message
 
 Edits a message spoken by the bot, by message ID
 """
+@utils.requires_admin
 async def edit(message):
-    # Only allow if user has correct permissions
-    roles = [x.id for x in message.author.roles]
-    if ADMIN_ACCESS not in roles:
-        return None
-
     try:
         payload = utils.remove_command(message.content)
         edit_id = utils.get_command(payload)
@@ -185,12 +177,8 @@ class CustomCommands:
 
     Input: message - Discord message object
     """
+    @utils.requires_admin
     async def define_cmd(self, message):
-        # Only allow if user has correct permissions
-        roles = [x.id for x in message.author.roles]
-        if ADMIN_ACCESS not in roles:
-            return None
-
         # First remove the "define" command
         new_cmd = utils.remove_command(message.content)
         # Then parse the new command
@@ -228,12 +216,8 @@ class CustomCommands:
 
     Input: message - Discord message object
     """
+    @utils.requires_admin
     async def remove_cmd(self, message):
-        # Only allow if user has correct permissions
-        roles = [x.id for x in message.author.roles]
-        if ADMIN_ACCESS not in roles:
-            return None
-
         # First remove the "define" command
         new_cmd = utils.remove_command(message.content)
         # Then parse the command to remove
