+++ src/utils.py
@@ -1,3 +1,5 @@
+from config import ADMIN_ACCESS
+
 """
 Strip prefix
 
@@ -30,3 +32,20 @@ Input: mes - message to modify - str
 def remove_command(mes):
     request = mes.split()[1:]
     return " ".join(request)
+
+"""
+Requires admin
+
+Wrapper function for commands that has them do nothing if the author doesn't have the admin role
+"""
+def requires_admin(func):
+    async def wrapper(*args, **kwargs):
+        # expect message to be the last argument
+        message = args[-1]
+
+        roles = [x.id for x in message.author.roles]
+        if ADMIN_ACCESS not in roles:
+            return None
+
+        return await func(*args, **kwargs)
+    return wrapper
